+++ packages/expo-file-system/src/ExpoFileSystem.ts
@@ -2,9 +2,9 @@ import { NativeModule, requireNativeModule } from 'expo-modules-core';
 
 import type { Directory } from './Directory';
 import type { File } from './File';
+import type { PickMultipleFilesOptions, PickSingleFileOptions } from './File.types';
 import type { DownloadOptions, DownloadProgress } from './NetworkTasks.types';
 import type { PathInfo } from './Paths.types';
-import type { PickMultipleFilesOptions, PickSingleFileOptions } from './Picker.types';
 import type {
   FileSystemDownloadTask,
   FileSystemUploadTask,
