+++ packages/expo-file-system/src/FileSystem.types.ts
@@ -3,4 +3,3 @@ export * from './File.types';
 export * from './FileSystemWatcher.types';
 export * from './NetworkTasks.types';
 export * from './Paths.types';
-export * from './Picker.types';
